+++ 003 Code/Frontend/src/components/Main/Chart.js
@@ -1,11 +1,13 @@
 import * as React from 'react';
-import Title from './Title';
+import Title from '../general/Title';
 import ApexCharts from 'react-apexcharts';
 import { useState, useEffect } from 'react';
-import { ConfigWithToken, ManagerBaseApi } from '../authConfig';
+import { ConfigWithToken, ManagerBaseApi } from '../../auth/authConfig';
 import axios from 'axios';
+import { useNavigate } from 'react-router-dom';
 
 export default function Chart() {
+  const navigate = useNavigate();
   const [predictitems, setPredictItems] = useState([]);
   const predictItemsArray = Object.entries(predictitems);
   const config = ConfigWithToken();
@@ -13,7 +15,10 @@ export default function Chart() {
   useEffect(() => {
     axios
       .get(`${ManagerBaseApi}/state/predict/food`, config)
-      .then((res) => setPredictItems(res.data));
+      .then((res) => setPredictItems(res.data))
+      .catch((err) => {
+        err.response.status === 401 && navigate('/');
+      });
   }, []);
   return (
     <React.Fragment>
@@ -26,8 +31,7 @@ export default function Chart() {
             data: predictItemsArray.map((items) => items[1]),
           },
         ]}
-        width={800}
-        height={170}
+        height={270}
         options={{
           chart: {
             toolbar: { show: false },
