--- 003 Code/Frontend/src/keycloaks/Protected.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { useKeycloak } from '@react-keycloak/web';
-
-const Protected = () => {
-  const { keycloak, initialized } = useKeycloak();
-  console.log(keycloak);
-  return <div>Protected</div>;
-};
-
-export default Protected;
