--- 003 Code/Frontend/src/keycloaks/Succes.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Succes = () => {
-  return <div>Succes</div>;
-};
-
-export default Succes;
