+++ 003 Code/Frontend/src/login/LoginFirst.js
@@ -3,23 +3,17 @@ import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
 import CssBaseline from '@mui/material/CssBaseline';
 import TextField from '@mui/material/TextField';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import Checkbox from '@mui/material/Checkbox';
-import Link from '@mui/material/Link';
 import Paper from '@mui/material/Paper';
 import Box from '@mui/material/Box';
 import Grid from '@mui/material/Grid';
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
 import Typography from '@mui/material/Typography';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
-import Copyright from '../components/Copyright';
-import background from '../assets/capstone_background.png';
+import Copyright from '../components/general/Copyright';
+import background from '../image/capstone_background.png';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
 import Skeleton from '@mui/material/Skeleton';
-import { ConfigWithToken, ManagerBaseApi } from '../authConfig';
-
-// TODO remove, this demo shouldn't need to reset the theme.
 
 const defaultTheme = createTheme();
